Server cleanup/Connecting to game client

+++ Player.java
@@ -3,19 +3,19 @@ import java.io.Serializable;
 public class Player implements Serializable{
     private static final long serialVersionUID = 13412431243L;
 
-    double x;
-    double y;
-    double rotation;
+    float x;
+    float y;
+    float rotation;
     int[] guns = new int[2];
     String name;
 
-    public Player(double x, double y, double rotation) {
+    public Player(float x, float y, float rotation) {
         this.x = x;
         this.y = y;
         this.rotation = rotation;
     }
 
-    public void setLocation (double[] location) {
+    public void setLocation (float[] location) {
         this.x = location[0];
         this.y = location[1];
     }
